requirements.txt
--extra-index-url https://mirrors.aliyun.com/pypi/simple

certifi==2020.12.5; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
cffi==1.14.4; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0"
chardet==3.0.4; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
click==7.1.2; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
colorama==0.4.4; python_version >= "3.5" and python_full_version < "3.0.0" and sys_platform == "win32" or sys_platform == "win32" and python_version >= "3.5" and python_full_version >= "3.5.0"
cryptography==3.3.1; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.6.0")
ecdsa==0.14.1; python_version >= "2.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0"
fastapi==0.62.0; python_version >= "3.6"
h11==0.11.0; python_version >= "3.6"
httpcore==0.12.2; python_version >= "3.6"
httpx==0.16.1; python_version >= "3.6"
idna==2.10; python_version >= "3.6" and python_full_version < "3.0.0" or python_version >= "3.6" and python_full_version >= "3.5.0"
jinja2==2.11.2; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.5.0")
loguru==0.5.3; python_version >= "3.5"
markupsafe==1.1.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
msal==1.7.0
pyasn1==0.4.8; python_version >= "3.5" and python_version < "4"
pycparser==2.20; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0"
pydantic==1.7.3; python_version >= "3.6"
pyjwt==1.7.1
python-jose==3.2.0
pyyaml==5.3.1; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.5.0")
redis==3.5.3; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.5.0")
requests==2.25.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
rfc3986==1.4.0; python_version >= "3.6"
rsa==4.6; python_version >= "3.5" and python_version < "4"
six==1.15.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0"
sniffio==1.2.0; python_version >= "3.6"
starlette==0.13.6; python_version >= "3.6"
typing-extensions==3.7.4.3; python_version < "3.8"
urllib3==1.26.2; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version < "4"
uvicorn==0.12.3
win32-setctime==1.0.3; sys_platform == "win32" and python_version >= "3.5"
azure-storage-blob==12.6.0
